Convert About pages to RST content in database
Removes all about hard-coded About content in favor of storing it in
the database via the admin site. As a side effect, introduces database
calls to (acceptably) load content on appropriate pages and
(unfortunately) render the About menu in the navbar. Addresses several
issues in Github: fixes #84, makes obsolete #24, #59, #83, and #86.